+++ rows/plugins/xlsx.py
@@ -33,7 +33,7 @@ def cell_to_python(cell):
     elif cell.number_format.lower() == "yyyy-mm-dd":
         return str(cell.value).split(" 00:00:00")[0]
     elif cell.number_format.lower() == "yyyy-mm-dd hh:mm:ss":
-        return str(cell.value)
+        return str(cell.value).split('.')[0]
     elif cell.number_format.endswith("%"):
         return "{}%".format(cell.value * 100)
     elif cell.value is None:
@@ -73,17 +73,39 @@ def import_from_xlsx(filename_or_fobj, sheet_name=None, sheet_index=0,
     return create_table([header] + all_rows, meta=metadata, *args, **kwargs)
 
 
-def _write_cell(sheet, row_index, col_index, value, field_type):
-    '''Write a cell to the sheet, fixing value/formatting if needed'''
+FORMATTING_STYLES = {
+        fields.DateField: 'YYYY-MM-DD',
+        fields.DatetimeField: 'YYYY-MM-DD HH:MM:SS',
+        fields.PercentField: '0.00%',
+}
 
-    cell = sheet.cell(row=row_index + 1, column=col_index + 1)
-    cell.value = value
 
-    if field_type is fields.PercentField:
-        cell.number_format = '0.00%'
-    elif field_type is fields.DatetimeField:
-        cell.value = str(cell.value).split(' 00:00:00')[0]
-        cell.number_format = 'YYYY-MM-DD'
+def _python_to_xlsx(field_types):
+
+    def convert_value(field_type, value):
+
+        number_format = FORMATTING_STYLES.get(field_type, None)
+
+        if field_type not in (
+                fields.BoolField,
+                fields.DateField,
+                fields.DatetimeField,
+                fields.DecimalField,
+                fields.FloatField,
+                fields.IntegerField,
+                fields.PercentField,
+                fields.TextField,
+        ):
+            # BinaryField, DatetimeField, JSONField or unknown
+            value = field_type.serialize(value)
+
+        return value, number_format
+
+    def convert_row(row):
+        return [convert_value(field_type, value)
+                for field_type, value in zip(field_types, row)]
+
+    return convert_row
 
 
 def export_to_xlsx(table, filename_or_fobj=None, sheet_name='Sheet1', *args,
@@ -92,21 +114,22 @@ def export_to_xlsx(table, filename_or_fobj=None, sheet_name='Sheet1', *args,
     workbook = Workbook()
     sheet = workbook.active
     sheet.title = sheet_name
-    field_names = list(enumerate(table.fields))
     prepared_table = prepare_to_export(table, *args, **kwargs)
 
     # Write header
-    header = prepared_table.next()
-    for col_index, field_name in enumerate(header):
-        _write_cell(sheet, 0, col_index, field_name, fields.TextField)
+    field_names = prepared_table.next()
+    for col_index, field_name in enumerate(field_names):
+        cell = sheet.cell(row=1, column=col_index + 1)
+        cell.value = field_name
 
     # Write sheet rows
-    table_fields = table.fields
+    _convert_row = _python_to_xlsx(map(table.fields.get, field_names))
     for row_index, row in enumerate(prepared_table, start=1):
-        for col_index, field_name in field_names:
-            _write_cell(sheet, row_index, col_index,
-                        value=row[col_index],
-                        field_type=table_fields[field_name])
+        for col_index, (value, number_format) in enumerate(_convert_row(row)):
+            cell = sheet.cell(row=row_index + 1, column=col_index + 1)
+            cell.value = value
+            if number_format is not None:
+                cell.number_format = number_format
 
     if filename_or_fobj is not None:
         _, fobj = get_filename_and_fobj(filename_or_fobj, mode='wb')
@@ -120,5 +143,3 @@ def export_to_xlsx(table, filename_or_fobj=None, sheet_name='Sheet1', *args,
         result = fobj.read()
         fobj.close()
         return result
-
-