+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 - [#170](https://github.com/turicas/rows/issues/170) (SQLite plugin) Error
   converting `int` and `float` when value is `None`
+- [#168](https://github.com/turicas/rows/issues/168) Use `Field.serialize`
+  if does not know the field type (affecting: XLS, XLSX and SQLite plugins)
 
 
 ## Version `0.2.0`